+++ requests_cache/policy/__init__.py
@@ -24,3 +24,4 @@ from .expiration import *
 from .settings import *
 from .directives import CacheDirectives, set_request_headers
 from .actions import CacheActions
+from .plan import RequestCachePlan